# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "BirdId": {
        "enum": [
          "vulture",
          "owl",
          "crow",
          "cock",
          "peacock"
        ],
        "title": "BirdId",
        "type": "string"
      },
      "BirdSelectionInput": {
        "properties": {
          "as_of_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of Date"
          },
          "as_of_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of Time"
          },
          "bird": {
            "$ref": "#/components/schemas/BirdId"
          },
          "iana_tz": {
            "title": "Iana Tz",
            "type": "string"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "location_name": {
            "title": "Location Name",
            "type": "string"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "method": {
            "const": "bird",
            "title": "Method",
            "type": "string"
          },
          "target_date": {
            "format": "date",
            "title": "Target Date",
            "type": "string"
          },
          "target_time": {
            "format": "time",
            "title": "Target Time",
            "type": "string"
          }
        },
        "required": [
          "target_date",
          "target_time",
          "location_name",
          "latitude",
          "longitude",
          "iana_tz",
          "method",
          "bird"
        ],
        "title": "BirdSelectionInput",
        "type": "object"
      },
      "BirdSelectionSummaryInput": {
        "properties": {
          "as_of_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of Date"
          },
          "as_of_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of Time"
          },
          "bird": {
            "$ref": "#/components/schemas/BirdId"
          },
          "days": {
            "default": 31,
            "maximum": 31.0,
            "minimum": 1.0,
            "title": "Days",
            "type": "integer"
          },
          "iana_tz": {
            "title": "Iana Tz",
            "type": "string"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "location_name": {
            "title": "Location Name",
            "type": "string"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "method": {
            "const": "bird",
            "title": "Method",
            "type": "string"
          },
          "min_effect": {
            "default": "good",
            "enum": [
              "good",
              "very_good"
            ],
            "title": "Min Effect",
            "type": "string"
          },
          "target_date": {
            "format": "date",
            "title": "Target Date",
            "type": "string"
          },
          "target_time": {
            "format": "time",
            "title": "Target Time",
            "type": "string"
          }
        },
        "required": [
          "target_date",
          "target_time",
          "location_name",
          "latitude",
          "longitude",
          "iana_tz",
          "method",
          "bird"
        ],
        "title": "BirdSelectionSummaryInput",
        "type": "object"
      },
...
      "BirthBirdResponse": {
        "properties": {
          "bharana_pakshi": {
            "$ref": "#/components/schemas/BirdId"
          },
          "birth_bird": {
            "$ref": "#/components/schemas/BirdId"
          },
          "padu_pakshi": {
            "$ref": "#/components/schemas/BirdId"
          }
        },
        "required": [
          "birth_bird",
          "padu_pakshi",
          "bharana_pakshi"
        ],
        "title": "BirthBirdResponse",
        "type": "object"
      },
      "BirthDateTimeInput": {
        "properties": {
          "as_of_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of Date"
          },
          "as_of_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of Time"
          },
          "birth_date": {
            "format": "date",
            "title": "Birth Date",
            "type": "string"
          },
          "birth_time": {
            "format": "time",
            "title": "Birth Time",
            "type": "string"
          },
          "iana_tz": {
            "title": "Iana Tz",
            "type": "string"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "location_name": {
            "title": "Location Name",
            "type": "string"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "method": {
            "const": "birth_datetime",
            "title": "Method",
            "type": "string"
          },
          "target_date": {
            "format": "date",
            "title": "Target Date",
            "type": "string"
          },
          "target_time": {
            "format": "time",
            "title": "Target Time",
            "type": "string"
          }
        },
        "required": [
          "target_date",
          "target_time",
          "location_name",
          "latitude",
          "longitude",
          "iana_tz",
          "method",
          "birth_date",
          "birth_time"
        ],
        "title": "BirthDateTimeInput",
        "type": "object"
      },
      "BirthDateTimeSummaryInput": {
        "properties": {
          "as_of_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of Date"
          },
          "as_of_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of Time"
          },
          "birth_date": {
            "format": "date",
            "title": "Birth Date",
            "type": "string"
          },
          "birth_time": {
            "format": "time",
            "title": "Birth Time",
            "type": "string"
          },
          "days": {
            "default": 31,
            "maximum": 31.0,
            "minimum": 1.0,
            "title": "Days",
            "type": "integer"
          },
          "iana_tz": {
            "title": "Iana Tz",
            "type": "string"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "location_name": {
            "title": "Location Name",
            "type": "string"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "method": {
            "const": "birth_datetime",
            "title": "Method",
            "type": "string"
          },
          "min_effect": {
            "default": "good",
            "enum": [
              "good",
              "very_good"
            ],
            "title": "Min Effect",
            "type": "string"
          },
          "target_date": {
            "format": "date",
            "title": "Target Date",
            "type": "string"
          },
          "target_time": {
            "format": "time",
            "title": "Target Time",
            "type": "string"
          }
        },
        "required": [
          "target_date",
          "target_time",
          "location_name",
          "latitude",
          "longitude",
          "iana_tz",
          "method",
          "birth_date",
          "birth_time"
        ],
        "title": "BirthDateTimeSummaryInput",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NakshatraPakshaInput": {
        "properties": {
          "as_of_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of Date"
          },
          "as_of_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of Time"
          },
          "iana_tz": {
            "title": "Iana Tz",
            "type": "string"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "location_name": {
            "title": "Location Name",
            "type": "string"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "method": {
            "const": "nakshatra_paksha",
            "title": "Method",
            "type": "string"
          },
          "moon_rashi_index": {
            "anyOf": [
              {
                "maximum": 12.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moon Rashi Index"
          },
          "nakshatra_index": {
            "title": "Nakshatra Index",
            "type": "integer"
          },
          "paksha": {
            "$ref": "#/components/schemas/PakshaId"
          },
          "target_date": {
            "format": "date",
            "title": "Target Date",
            "type": "string"
          },
          "target_time": {
            "format": "time",
            "title": "Target Time",
            "type": "string"
          }
        },
        "required": [
          "target_date",
          "target_time",
          "location_name",
          "latitude",
          "longitude",
          "iana_tz",
          "method",
          "nakshatra_index",
          "paksha"
        ],
        "title": "NakshatraPakshaInput",
        "type": "object"
      },
      "NakshatraPakshaSummaryInput": {
        "properties": {
          "as_of_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of Date"
          },
          "as_of_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of Time"
          },
          "days": {
            "default": 31,
            "maximum": 31.0,
            "minimum": 1.0,
            "title": "Days",
            "type": "integer"
          },
          "iana_tz": {
            "title": "Iana Tz",
            "type": "string"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "location_name": {
            "title": "Location Name",
            "type": "string"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "method": {
            "const": "nakshatra_paksha",
            "title": "Method",
            "type": "string"
          },
          "min_effect": {
            "default": "good",
            "enum": [
              "good",
              "very_good"
            ],
            "title": "Min Effect",
            "type": "string"
          },
          "moon_rashi_index": {
            "anyOf": [
              {
                "maximum": 12.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moon Rashi Index"
          },
          "nakshatra_index": {
            "title": "Nakshatra Index",
            "type": "integer"
          },
          "paksha": {
            "$ref": "#/components/schemas/PakshaId"
          },
          "target_date": {
            "format": "date",
            "title": "Target Date",
            "type": "string"
          },
          "target_time": {
            "format": "time",
            "title": "Target Time",
            "type": "string"
          }
        },
        "required": [
          "target_date",
          "target_time",
          "location_name",
          "latitude",
          "longitude",
          "iana_tz",
          "method",
          "nakshatra_index",
          "paksha"
        ],
        "title": "NakshatraPakshaSummaryInput",
        "type": "object"
      },
...
      "PakshaId": {
        "enum": [
          "waxing",
          "waning"
        ],
        "title": "PakshaId",
        "type": "string"
      },
...
      "ScheduleResponse": {
        "properties": {
          "bharana_pakshi": {
            "$ref": "#/components/schemas/BirdId"
          },
          "birth_bird": {
            "$ref": "#/components/schemas/BirdId"
          },
          "chandrashtama": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChandrashtamaWindow"
              },
              {
                "type": "null"
              }
            ]
          },
          "current_period": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SubPeriod"
              },
              {
                "type": "null"
              }
            ]
          },
          "disha_shool": {
            "title": "Disha Shool",
            "type": "string"
          },
          "engine": {
            "$ref": "#/components/schemas/EngineMetadata"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "major_periods": {
            "items": {
              "$ref": "#/components/schemas/MajorPeriod"
            },
            "title": "Major Periods",
            "type": "array"
          },
          "next_period": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SubPeriod"
              },
              {
                "type": "null"
              }
            ]
          },
          "next_sunrise": {
            "format": "date-time",
            "title": "Next Sunrise",
            "type": "string"
          },
          "padu_pakshi": {
            "$ref": "#/components/schemas/BirdId"
          },
          "paksha": {
            "$ref": "#/components/schemas/PakshaId"
          },
          "summary": {
            "$ref": "#/components/schemas/ScheduleSummary"
          },
          "sunrise": {
            "format": "date-time",
            "title": "Sunrise",
            "type": "string"
          },
          "sunset": {
            "format": "date-time",
            "title": "Sunset",
            "type": "string"
          },
          "tara_bala": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaraBala"
              },
              {
                "type": "null"
              }
            ]
          },
          "weekday": {
            "$ref": "#/components/schemas/WeekdayId"
          }
        },
        "required": [
          "engine",
          "location",
          "sunrise",
          "sunset",
          "next_sunrise",
          "birth_bird",
          "tara_bala",
          "chandrashtama",
          "disha_shool",
          "paksha",
          "weekday",
          "padu_pakshi",
          "bharana_pakshi",
          "current_period",
          "next_period",
          "major_periods",
          "summary"
        ],
        "title": "ScheduleResponse",
        "type": "object"
      },
...
      "WeekdayId": {
        "enum": [
          "sunday",
          "monday",
          "tuesday",
          "wednesday",
          "thursday",
          "friday",
          "saturday"
        ],
        "title": "WeekdayId",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Fernando Family Astrology API",
    "title": "Fernando Family Astrology",
    "version": "4.8.7"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/pancha-pakshi/birth-bird": {
      "post": {
        "operationId": "birth_bird_api_v1_pancha_pakshi_birth_bird_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "birth_datetime": "#/components/schemas/BirthDateTimeInput",
                    "nakshatra_paksha": "#/components/schemas/NakshatraPakshaInput"
                  },
                  "propertyName": "method"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/BirthDateTimeInput"
                  },
                  {
                    "$ref": "#/components/schemas/NakshatraPakshaInput"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BirthBirdResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Birth Bird",
        "tags": [
          "pancha-pakshi"
        ]
      }
    },
    "/api/v1/pancha-pakshi/current": {
      "post": {
        "operationId": "current_api_v1_pancha_pakshi_current_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "bird": "#/components/schemas/BirdSelectionInput",
                    "birth_datetime": "#/components/schemas/BirthDateTimeInput",
                    "nakshatra_paksha": "#/components/schemas/NakshatraPakshaInput"
                  },
                  "propertyName": "method"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/BirthDateTimeInput"
                  },
                  {
                    "$ref": "#/components/schemas/NakshatraPakshaInput"
                  },
                  {
                    "$ref": "#/components/schemas/BirdSelectionInput"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Current Api V1 Pancha Pakshi Current Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Current",
        "tags": [
          "pancha-pakshi"
        ]
      }
    },
...
    "/api/v1/pancha-pakshi/schedule": {
      "post": {
        "operationId": "schedule_api_v1_pancha_pakshi_schedule_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "bird": "#/components/schemas/BirdSelectionInput",
                    "birth_datetime": "#/components/schemas/BirthDateTimeInput",
                    "nakshatra_paksha": "#/components/schemas/NakshatraPakshaInput"
                  },
                  "propertyName": "method"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/BirthDateTimeInput"
                  },
                  {
                    "$ref": "#/components/schemas/NakshatraPakshaInput"
                  },
                  {
                    "$ref": "#/components/schemas/BirdSelectionInput"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Schedule",
        "tags": [
          "pancha-pakshi"
        ]
      }
    },
    "/api/v1/pancha-pakshi/summary": {
      "post": {
        "operationId": "summary_api_v1_pancha_pakshi_summary_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "bird": "#/components/schemas/BirdSelectionSummaryInput",
                    "birth_datetime": "#/components/schemas/BirthDateTimeSummaryInput",
                    "nakshatra_paksha": "#/components/schemas/NakshatraPakshaSummaryInput"
                  },
                  "propertyName": "method"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/BirthDateTimeSummaryInput"
                  },
                  {
                    "$ref": "#/components/schemas/NakshatraPakshaSummaryInput"
                  },
                  {
                    "$ref": "#/components/schemas/BirdSelectionSummaryInput"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Summary Api V1 Pancha Pakshi Summary Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Summary",
        "tags": [
          "pancha-pakshi"
        ]
      }
    },
    "/api/v1/pancha-pakshi/windows": {
      "post": {
        "operationId": "windows_api_v1_pancha_pakshi_windows_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "bird": "#/components/schemas/BirdSelectionWindowsInput",
                    "birth_datetime": "#/components/schemas/BirthDateTimeWindowsInput",
                    "nakshatra_paksha": "#/components/schemas/NakshatraPakshaWindowsInput"
                  },
                  "propertyName": "method"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/BirthDateTimeWindowsInput"
                  },
                  {
                    "$ref": "#/components/schemas/NakshatraPakshaWindowsInput"
                  },
                  {
                    "$ref": "#/components/schemas/BirdSelectionWindowsInput"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Windows Api V1 Pancha Pakshi Windows Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Windows",
        "tags": [
          "pancha-pakshi"
        ]
      }
    },
...
  }
}